avalanche queries
