+++ interchaintest/tests/devlog.txt
@@ -1,0 +1,8 @@
+- It doesn't seem like CIDT's interchaintest code works in the most up to date branch, avalanche-support in github.com/ConsiderItDone/interchaintest
+   - I wanted to use their code to help launch the docker containers for Avalanche
+- Writing code for Avalanche to deploy contracts. Subnet-evm uses avalanchego v1.10.1, so naturally I tried using v1.10.1 in this repo as well.
+   - I got the build error /home/kyle/go/pkg/mod/github.com/ava-labs/avalanchego@v1.10.1/database/leveldb/db.go:228:3: unknown field MaxManifestFileSize in struct literal of type opt.Options (exit status 1)
+   - Subnet-evm (actually, avalanchego) uses a different version of goleveldb than cosmos-sdk has pinned in the go.mod replace statements.
+   - For the time being, I tried removing the dependency on subnet-evm and upgrading to avalanchego v1.10.10.
+      - wallet.NewWalletFromURI(ctx, nodeURI, kc) now has a compile error since the Avalanche team removed that function.
+      - I am not sure what they replaced it with, the other wallet functions make no sense.
